Handle missing args from long flags that end `=`
Adjust `Token::LongEq` to contain the value after the `=` (which may be the
empty string) so the parser has access (instead of saving it in `pending`).

In the parser, if the flag is boolean and there is an arg then this is an
error, exit with an error message.

Otherwise, take the bundled argument, create a tokenizer for it, and call
the flag's parser to process it.

So,

`--foo=` when `foo` is boolean is an error

`--foo=` when `foo` takes some other value is parsed as the empty string,
which may or may not be a valid value for that flag type.

Add a test for this behaviour.

Change `new_with_iterator()` to *not* skip the first argument of the iterator,
carry that out in `new()`. Adjust the tests accordingly.